TEA-18: Refacotor - Multiple model system - 4
* Code review fix
* Small updates about member/namespace names

+++ Teapot/src/Models/ModelManager.cpp
@@ -56,7 +56,7 @@ namespace Teapot
 	{
 		for (const auto& modelVec : s_AllModels)
 		{
-			unsigned int modelVecIdx{};
+			size_t modelVecIdx{};
 			for (const auto& model : *modelVec)
 			{
 				if (model == GetSelectedModel())
